unified_server/requirements.txt
# Unified Posture Analysis Server - Dependencies
# Team 2 - Consolidated System

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
asyncpg>=0.29.0
alembic>=1.14.0

# Authentication
pyjwt>=2.9.0

# AI Integration
groq>=0.11.0

# HTTP Client
httpx>=0.27.0
requests>=2.32.0

# Configuration
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# Utilities
python-multipart>=0.0.12
